feat:external sharing data format modify
